# application-config-service/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Application Config Service API
  description: |
    RESTful API for managing application configurations and modules in the Onified platform.
    
    This service provides endpoints for:
    - Creating, reading, updating, and deleting applications
    - Managing application modules
    - Application configuration management
    
    ## Features
    - **Application Management**: Full CRUD operations for applications
    - **Module Management**: Create and manage modules within applications
    - **Active/Inactive Status**: Toggle application and module status
    - **Standardized Responses**: Consistent API response format
    
    ## Authentication
    This service requires proper authentication and authorization.
    
    ## Rate Limiting
    API calls are subject to rate limiting to ensure service stability.
  version: 1.0.0
  contact:
    name: Onified Development Team
    email: dev@onified.ai
    url: https://onified.ai
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:9082
    description: Local Development Server
  - url: https://api.onified.ai
    description: Production Server
...
paths:
  /api/applications:
    post:
      tags:
        - Application Management
      summary: Create a new application
      description: Creates a new application with the provided details. The application code must be unique.
      operationId: createApplication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationRequestDTO'
            examples:
              create-app:
                summary: Create Application Example
                value:
                  appCode: "APP001"
                  displayName: "User Management System"
                  isActive: true
      responses:
        '201':
          description: Application created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ApplicationResponseDTO'
              examples:
                success:
                  summary: Success Response
                  value:
                    statusCode: 201
                    status: "SUCCESS"
                    body:
                      appCode: "APP001"
                      displayName: "User Management System"
                      isActive: true
        '400':
          description: Bad request - Invalid input data
        '409':
          description: Conflict - Application code already exists
        '500':
          description: Internal server error
    get:
      tags:
        - Application Management
      summary: Get all applications
      description: Retrieves a list of all applications in the system.
      operationId: getAllApplications
      responses:
        '200':
          description: Applications retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_List_ApplicationResponseDTO'
        '500':
          description: Internal server error
  /api/applications/{appCode}:
    get:
      tags:
        - Application Management
      summary: Get application by code
      description: Retrieves an application by its unique application code.
      operationId: getApplicationByAppCode
      parameters:
        - name: appCode
          in: path
          required: true
          description: Unique application code
          schema:
            type: string
            example: "APP001"
      responses:
        '200':
          description: Application found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ApplicationResponseDTO'
        '404':
          description: Application not found
        '500':
          description: Internal server error
    put:
      tags:
        - Application Management
      summary: Update an application
      description: Updates an existing application with new details. The application code in the path must match the one in the request body.
      operationId: updateApplication
      parameters:
        - name: appCode
          in: path
          required: true
          description: Application code to update
          schema:
            type: string
            example: "APP001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationRequestDTO'
            examples:
              update-app:
                summary: Update Application Example
                value:
                  appCode: "APP001"
                  displayName: "Updated User Management System"
                  isActive: false
      responses:
        '200':
          description: Application updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ApplicationResponseDTO'
        '400':
          description: Bad request - Invalid input data
        '404':
          description: Application not found
        '500':
          description: Internal server error
    delete:
      tags:
        - Application Management
      summary: Delete an application
      description: Deletes an application by its application code. This operation is irreversible.
      operationId: deleteApplication
      parameters:
        - name: appCode
          in: path
          required: true
          description: Application code to delete
          schema:
            type: string
            example: "APP001"
      responses:
        '204':
          description: Application deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_String'
              examples:
                success:
                  summary: Success Response
                  value:
                    statusCode: 204
                    status: "SUCCESS"
                    body: "Application deleted successfully."
        '404':
          description: Application not found
        '500':
          description: Internal server error
...
components:
  schemas:
    ApplicationRequestDTO:
      type: object
      description: Request DTO for creating or updating an application
      required:
        - appCode
        - displayName
        - isActive
      properties:
        appCode:
          type: string
          description: Unique application code identifier
          example: "APP001"
          minLength: 1
          maxLength: 50
        displayName:
          type: string
          description: Display name for the application
          example: "User Management System"
          minLength: 1
          maxLength: 100
        isActive:
          type: boolean
          description: Whether the application is active
          example: true
    ApplicationResponseDTO:
      type: object
      description: Response DTO for application data
      properties:
        appCode:
          type: string
          description: Unique application code identifier
          example: "APP001"
        displayName:
          type: string
          description: Display name for the application
          example: "User Management System"
        isActive:
          type: boolean
          description: Whether the application is active
          example: true
...
    ApiResponse_ApplicationResponseDTO:
      type: object
      description: Standardized API response wrapper for ApplicationResponseDTO
      properties:
        statusCode:
          type: integer
          description: HTTP status code
          example: 200
        status:
          type: string
          description: Status message
          example: "SUCCESS"
        body:
          $ref: '#/components/schemas/ApplicationResponseDTO'
...
    ApiResponse_List_ApplicationResponseDTO:
      type: object
      description: Standardized API response wrapper for list of ApplicationResponseDTO
      properties:
        statusCode:
          type: integer
          description: HTTP status code
          example: 200
        status:
          type: string
          description: Status message
          example: "SUCCESS"
        body:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationResponseDTO'
...
    ApiResponse_String:
      type: object
      description: Standardized API response wrapper for String
      properties:
        statusCode:
          type: integer
          description: HTTP status code
          example: 204
        status:
          type: string
          description: Status message
          example: "SUCCESS"
        body:
          type: string
          description: Response body containing the actual data
          example: "Application deleted successfully."
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
security:
  - bearerAuth: [] 